Remove create/* folder for API, include priv methods
