+++ packages/api/src/index.ts
@@ -2,11 +2,21 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import { ProviderInterface } from '@polkadot/api-provider/types';
-import { ApiInterface, ApiInterface$Section } from './types';
+import { ProviderInterface, ProviderInterface$Callback } from '@polkadot/api-provider/types';
+import { Interfaces, Interface$Sections } from '@polkadot/jsonrpc/types';
+import { SectionItem } from '@polkadot/params/types';
+import { Storages } from '@polkadot/storage/types';
+import { ApiInterface, ApiInterface$Section, ApiInterface$Section$Method } from './types';
+
+import formatInputs from '@polkadot/api-format/input';
+import formatOutput from '@polkadot/api-format/output';
+import interfaces from '@polkadot/jsonrpc/index';
+import decodeParams from '@polkadot/params/decode';
+import signature from '@polkadot/params/signature';
 import assert from '@polkadot/util/assert';
+import ExtError from '@polkadot/util/ext/error';
 import isFunction from '@polkadot/util/is/function';
-import createInterface from './create/interface';
+import isUndefined from '@polkadot/util/is/undefined';
 
 /**
  * @example
@@ -14,6 +24,7 @@ import createInterface from './create/interface';
  *
  * import Api from '@polkadot/api';
  * import WsProvider from '@polkadot/api-provider/ws';
+ *
  * const provider = new WsProvider('http://127.0.0.1:9944');
  * const api = new Api(provider);
  * ```
@@ -35,9 +46,100 @@ export default class Api implements ApiInterface {
 
     this._provider = provider;
 
-    this.author = createInterface(this._provider, 'author');
-    this.chain = createInterface(this._provider, 'chain');
-    this.state = createInterface(this._provider, 'state');
-    this.system = createInterface(this._provider, 'system');
+    this.author = this.createInterface('author');
+    this.chain = this.createInterface('chain');
+    this.state = this.createInterface('state');
+    this.system = this.createInterface('system');
+  }
+
+  private createInterface (section: Interface$Sections): ApiInterface$Section {
+    const definition = interfaces[section];
+
+    assert(!isUndefined(definition), `Unable to find section '${section}`);
+
+    // @ts-ignore undefined check done in assert
+    const methods = definition.public;
+
+    return Object
+      .keys(methods)
+      .reduce((exposed, name) => {
+        const rpcName = `${section}_${name}`;
+        const def = methods[name];
+
+        exposed[name] = def.isSubscription
+          ? this.createMethodSubscribe(rpcName, def)
+          : this.createMethodSend(rpcName, def);
+
+        return exposed;
+      }, {} as ApiInterface$Section);
+  }
+
+  private createMethodSend (rpcName: string, method: SectionItem<Interfaces>): ApiInterface$Section$Method {
+    const call = async (...values: Array<any>): Promise<any> => {
+      // TODO Warn on deprecated methods
+      try {
+        const params = formatInputs(method.params, values);
+        const result = await this._provider.send(rpcName, params);
+
+        return this.formatResult(method, params, values, result);
+      } catch (error) {
+        throw new ExtError(`${signature(method)}:: ${error.message}`, (error as ExtError).code, undefined);
+      }
+    };
+
+    return call as ApiInterface$Section$Method;
+  }
+
+  private createMethodSubscribe (rpcName: string, method: SectionItem<Interfaces>): ApiInterface$Section$Method {
+    const unsubscribe = (subscriptionId: any): Promise<any> =>
+      this._provider.unsubscribe(rpcName, method.subscribe[1], subscriptionId);
+    const _call = async (...values: Array<any>): Promise<any> => {
+      try {
+        const cb: ProviderInterface$Callback = values.pop();
+
+        assert(isFunction(cb), `Expected callback in last position of params`);
+
+        const params = formatInputs(method.params, values);
+        const update = (error: Error | null, result?: any) => {
+          cb(error, this.formatResult(method, params, values, result));
+        };
+
+        return this._provider.subscribe(rpcName, method.subscribe[0], params, update);
+      } catch (error) {
+        throw new ExtError(`${signature(method)}:: ${error.message}`, (error as ExtError).code, undefined);
+      }
+    };
+
+    const call = _call as ApiInterface$Section$Method;
+
+    call.unsubscribe = unsubscribe;
+
+    return call;
+  }
+
+  private formatResult (method: SectionItem<Interfaces>, params: Array<any>, inputs: Array<any>, result?: any): any {
+    if (method.type === 'StorageResult') {
+      return this.formatStorageOutput(inputs[0][0], result);
+    }
+
+    if (method.type === 'StorageResultSet') {
+      return params[0].map((key: string, index: number) => {
+        const input = inputs[0][index][0];
+        const { changes = [] }: { block: string, changes: Array<[string, string]> } = result || {};
+        const value = changes.find(([_key]) => key === _key);
+
+        if (!value) {
+          return undefined;
+        }
+
+        return this.formatStorageOutput(input, value[1]);
+      });
+    }
+
+    return formatOutput(method.type, result);
+  }
+
+  private formatStorageOutput (key: SectionItem<Storages>, result?: any): any {
+    return decodeParams(key.type, result, 'latest', true).value;
   }
 }
